improved xp calculation

+++ src/main/java/fab/keepinventorypenalty/config/ConfigData.java
@@ -3,15 +3,15 @@ package fab.keepinventorypenalty.config;
 public class ConfigData
 {
     public int VERSION_DONT_CHANGE;
-    public int LossPercentage;
+    public float LossPercentage;
     public boolean PenaltyEnabled;
     public boolean RandomizePenalty;
-    public int RandomMin;
-    public int RandomMax;
+    public float RandomMin;
+    public float RandomMax;
     public boolean GlobalShame;
 
-    public ConfigData(int version, int lossPercentage, boolean penaltyEnabled, boolean randomizePenalty, int randomMin,
-                      int randomMax, boolean globalShame)
+    public ConfigData(int version, float lossPercentage, boolean penaltyEnabled, boolean randomizePenalty, float randomMin,
+                      float randomMax, boolean globalShame)
     {
         VERSION_DONT_CHANGE = version;
         LossPercentage = lossPercentage;
